+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug fixes
 
+* [#7766](https://github.com/rubocop-hq/rubocop/issues/7766): Rename rescue body vars when renaming exception name. ([@asterite][])
 * [#9342](https://github.com/rubocop-hq/rubocop/issues/9342): Fix an error for `Lint/RedundantDirGlobSort` when using `collection.sort`. ([@koic][])
 * [#9304](https://github.com/rubocop-hq/rubocop/issues/9304): Do not register an offense for `Style/ExplicitBlockArgument` when the `yield` arguments are not an exact match with the block arguments. ([@dvandersluis][])
 * [#8281](https://github.com/rubocop-hq/rubocop/issues/8281): Fix Style/WhileUntilModifier handling comments and assignment when correcting to modifier form. ([@Darhazer][])
@@ -5367,3 +5368,4 @@
 [@magneland]: https://github.com/magneland
 [@k-karen]: https://github.com/k-karen
 [@uplus]: https://github.com/uplus
+[@asterite]: https://github.com/asterite